apply_mapshaper_commands {rmapshaper}	R Documentation

Apply a mapshaper command string to a geojson object

Description

Apply a mapshaper command string to a geojson object

Usage

apply_mapshaper_commands(data, command, force_FC, sys = FALSE)

Arguments

data
geojson object or path to geojson file. If a file path, sys must be true

command
valid mapshaper command string

force_FC
should the output be forced to be a FeatureCollection (or Spatial*DataFrame) even if there are no attributes? Default TRUE. FeatureCollections are more compatible with rgdal::readOGR and geojsonio::geojson_sp. If FALSE and there are no attributes associated with the geometries, a GeometryCollection (or Spatial object with no dataframe) will be output.

sys
Should the system mapshaper be used instead of the bundled mapshaper? Gives better performance on large files. Requires the mapshaper node package to be installed and on the PATH.

Value

geojson

[Package rmapshaper version 0.4.3 Index]